% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributionVisualizations.R
\name{vnorm}
\alias{vnorm}
\title{Add together two numbers.}
\usage{
vnorm(a = -3, b = 3, mean = 0, sd = 1, lower.tail = TRUE,
  log.p = FALSE, fill = "red")
}
\arguments{
\item{a}{left-endpoint}

\item{b}{right-endpoint}
}
\value{
ggplot 2 graph object
}
\description{
Add together two numbers.
}
\examples{
vnorm(b = 1)
vnorm(a = 1.5, fill = "blue", lower.tail = FALSE)
}
